pyproject.toml
[tool.poetry]
name = "pointsbot"
version = "0.1.0"
description = ""
authors = ["NicholasBaldwinSE <nicholasbaldwinse@proton.me>"]
license = "LGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "2.3.2"
python-dotenv = "1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
